Allow using sqlite for testing.  Set conftest.{BIND,DBNAME}.
Several changes: some postgres-specific DB ops are conditioned on
using postgres or provided with sqlite options.

Also, update to different method of testing with known time values,
which is necessary for testing with sqlite.  Patching datetime.datetime
won't work with the sqlite driver, which makes use of isinstance()
to determine that a field is a datetime).

Mostly follows the pattern of:

1) remove the monkeypatching of datetime.datetime in favor of a
fixture "no_update_timestamp" that disables the sqlalchemy listener
functions, which had been used to update the timestamp for a Repo
or Binary at commit time, and

2) remove the "now" parameter from purge_repos(), as it's no longer
necessary

Exceptions: the two tests that now have "but_keep_new" in their names
fix logic errors in the original tests

Signed-off-by: Dan Mick <dmick@redhat.com>