Allow 'raw' repositories (simple directories of binaries)
These are useful for, say, Windows builds which have no package
manager.  A 'raw' type can be (less-safely)  inferred if the binary
tested doesn't match any of the known extensions for Debian or RPM
repos.  Safer is to use the create operation with a JSON post to
/repos/<project>/<ref>/<distro>/<distrover>/ and a JSON
payload including '{"type":"raw"}'.  (Types deb and rpm can also
be set on create.)

Raw repos are available immediately on POST to 'update'; no async
operations need be started.  Status is sent to shaman as usual, just
immediately on 'update'.

Signed-off-by: Dan Mick <dmick@redhat.com>