various places: fix 'remove end part of path'
The existing code was splitting on the 'binary name' string,
but that breaks if the binary name also appears somewhere
besides the end of the path (the case noticed was a branch with
'cephadm' in its name that also was uploading a binary named
'cephadm').  Remove the path only if it appears as a suffix.

Also, add a couple of test cases.

Signed-off-by: Dan Mick <dan.mick@redhat.com>